Reduce the overhead of extruder-inject.
Loading the mail gem causes a huge amount of time to be spent loading MIME
types.  Since we're not actually processing any mail, only writing it to a
file, avoiding loading it is much better for performance.  Move the parts
that actually need the mail gem into a separate file.  This brings the time
to inject a mail down from 2.0 seconds to 0.7 seconds, a 65% improvement.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>