Properly detect origin servers past the first.
If the origin server was not the first in the list, the algorithm would
incorrectly return nil, indicating one was not detected, instead of the
correct value.  Refactor the code to properly implement the intended
algorithm and detect origin servers that are not the first.  Add tests to
ensure that the behavior is as expected.

This change also has the consequence of short-circuiting the detection code,
potentially resulting in better performance.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>