Read messages from queue as binary.
The default encoding used by Ruby for files is that of the current
locale, which on most systems is UTF-8.  However, many mails are not
UTF-8, and the mail gem becomes unhappy if it gets strings which have
the wrong encoding.  Read files as binary to ensure that the mail gem
can parse them correctly.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>